refactoring to have pages directory.
