+++ src/pages/ExercisePage.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import CategoryList from "./CategoryList/CategoryList";
+import CategoryList from "../components/CategoryList/CategoryList";
 export default {
   name: "ExercisePage",
   props: ["exerciseId"],
